+++ tests/map_conversion/opendrive/conversion/plane_elements/test_plane.py
@@ -1,4 +1,5 @@
 import unittest
+from pyproj.transformer import CRS
 from crdesigner.map_conversion.opendrive.opendrive_conversion.plane_elements.plane import *
 from crdesigner.map_conversion.opendrive.opendrive_parser.elements import roadLanes
 from crdesigner.map_conversion.opendrive.opendrive_parser.elements.roadPlanView import PlanView
@@ -215,6 +216,47 @@ class TestParametricLane(unittest.TestCase):
         self.assertListEqual([], left.tolist())
         self.assertListEqual([], right.tolist())
 
+    def test_calc_vertices_transformed(self):
+        crs_from = CRS("+proj=tmerc +a=6378137 +b=6378137 +lon_0=11.663999726157273 +x_0=-0 "
+                       "+y_0=-5372577.5788830593 +k=1.0 +units=m +nadgrids=@null +wktext +no_defs")
+        crs_to = CRS("+proj=utm +zone=32 +ellps=WGS84")
+        transformer = Transformer.from_proj(crs_from, crs_to)
+        parametric_lane = self.init_lane()
+        parametric_lane.border_group.outer_border_offset = 0.0
+        parametric_lane.border_group.inner_border.width_coefficients.clear()
+        parametric_lane.border_group.inner_border.width_coefficients.append([5, 0, 0, 0])
+        # test on straight-line geometry of the initialised lane
+        # function should create three vertices: start, middle and end
+        parametric_lane.length = 46.911
+        left, right = parametric_lane.calc_vertices(error_tolerance=0.2, min_delta_s=0.3, transformer=transformer)
+
+        # test for calculating the vertices without sampling
+        self.assertEqual(94, len(left))
+        self.assertEqual(94, len(right))
+        # test for calculating the vertices with sampling
+        # self.assertEqual(3, len(left))
+        # self.assertEqual(3, len(right))
+
+        # test that inner border vertices are offset (in t-direction) by 5
+        # test for calculating the vertices without sampling
+        np.testing.assert_almost_equal([697725.0288082, 5348958.53774993], left[0], 3)
+        np.testing.assert_almost_equal([697757.153601, 5348992.87078342], left[-1], 3)
+        # test for calculating the vertices with sampling
+        # np.testing.assert_almost_equal([[22.574, 29.645], [39.161, 46.229], [55.806, 62.877]], left, 3)
+
+        # test that outer border vertices are offset (in t-direction) by -1
+        # test for calculating the vertices without sampling
+        np.testing.assert_almost_equal([697729.42382314, 5348954.44912354], right[0], 3)
+        np.testing.assert_almost_equal([697761.54861731, 5348988.78215698], right[-1], 3)
+        # test for calculating the vertices with sampling
+        # np.testing.assert_almost_equal([[26.817, 25.402], [43.404, 41.986], [60.049, 58.634]], right, 3)
+
+        # test that parametric lane with length < 0 returns empty lists
+        parametric_lane.length = -1
+        left, right = parametric_lane.calc_vertices(error_tolerance=0.2, min_delta_s=0.3, transformer=transformer)
+        self.assertListEqual([], left.tolist())
+        self.assertListEqual([], right.tolist())
+
     def test_zero_width_change_positions(self):
         lane = self.init_lane()
         lane.length = 30
@@ -234,7 +276,7 @@ class TestParametricLane(unittest.TestCase):
         lane.border_group.outer_border.width_coefficient_offsets.append(0)
         lane.border_group.outer_border.width_coefficient_offsets.append(30)
         roots = lane.zero_width_change_positions()
-        true_roots = P.polyroots(P.polyder(lane.border_group.outer_border.width_coefficients[0]))
+        true_roots = polynomial.polyroots(polynomial.polyder(lane.border_group.outer_border.width_coefficients[0]))
         np.testing.assert_almost_equal(true_roots, roots, 3)
         # test that reverse works
         lane.reverse = True