+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Versions of third-party packages
 - Use positive IDs for the cr2lanelet conversion
 - Structure for config parameters (similar as in commonroad-io)
+- Add version to lanelet2 xml elements
 
 ### Fixed
 - Lanelet2 projection